+++ integration-tests/src/eth.rs
@@ -7,76 +7,13 @@ use alloy::signers::local::PrivateKeySigner;
 use alloy::signers::Signer;
 use alloy::sol_types::{SolEvent, SolValue};
 use anyhow::{Context, Result};
+use mpc_node::indexer_eth::abi::{ChainSignaturesConstructor, SignatureRequestedEncoding};
 use mpc_primitives::LATEST_MPC_KEY_VERSION;
 use serde_json::Value;
 use std::time::Duration;
 
-pub use ChainSignatures::SignRequest;
-pub use ChainSignatures::SignatureResponded;
-
-alloy::sol! {
-    #[sol(rpc)]
-    contract ChainSignatures {
-        struct SignRequest {
-            bytes32 payload;
-            string path;
-            uint32 keyVersion;
-            string algo;
-            string dest;
-            string params;
-        }
-
-        struct AffinePoint {
-            uint256 x;
-            uint256 y;
-        }
-
-        struct Signature {
-            AffinePoint bigR;
-            uint256 s;
-            uint8 recoveryId;
-        }
-
-        function sign(SignRequest memory _request) external payable;
-        function getSignatureDeposit() external view returns (uint256);
-
-        event SignatureRequested(
-            address sender,
-            bytes32 payload,
-            uint32 keyVersion,
-            uint256 deposit,
-            uint256 chainId,
-            string path,
-            string algo,
-            string dest,
-            string params
-        );
-
-        event SignatureResponded(
-            bytes32 indexed requestId,
-            address responder,
-            Signature signature
-        );
-    }
-
-    // Event encoding used to derive the off-chain request id
-    event SignatureRequestedEncoding(
-        address sender,
-        bytes payload,
-        string path,
-        uint32 keyVersion,
-        uint256 chainId,
-        string algo,
-        string dest,
-        string params
-    );
-
-    // Constructor args for deployment
-    struct ChainSignaturesConstructor {
-        address mpcNetwork;
-        uint256 signatureDeposit;
-    }
-}
+pub use mpc_node::indexer_eth::abi::ChainSignatures;
+pub use mpc_node::indexer_eth::abi::ChainSignatures::{SignRequest, SignatureResponded};
 
 pub fn client(
     endpoint: &str,
